notebook/requirements.txt
pyspark==3.3
dash
plotly
kafka-python
vnstock
redis
pandas
numpy
